fix(android): cast compressionLevel to int in switch statement
Java does not support switch on double values. Casting to int is safe
since all valid compression level values (-1, 0-9) are whole numbers
passed from JS.

Fixes #341

+++ android/src/main/java/com/rnziparchive/RNZipArchiveModule.java
@@ -470,7 +470,7 @@ public class RNZipArchiveModule extends ReactContextBaseJavaModule {
   }
 
   private static CompressionLevel getCompressionLevel(double compressionLevel) {
-    switch (compressionLevel) {
+    switch ((int) compressionLevel) {
       case -1:
         return CompressionLevel.NORMAL;
       case 0:
